19/02 v1
Former-commit-id: 448be1db35d44805e98871f24ccf047b6ff3d4a8

+++ DrugSchedule.Services/Utils/DateAndTimeUtility.cs
@@ -23,7 +23,7 @@ internal static class DateUtility
 
         while (startDate <= endDate)
         {
-            if (((int)dayOfWeek & (int)startDate.DayOfWeek) == (int)startDate.DayOfWeek)
+            if (((byte)dayOfWeek & (byte)startDate.DayOfWeek) == (byte)startDate.DayOfWeek)
             {
                 yield return startDate;
             }
